Moved Home Work projects to initial work space
