+++ MentoAD/HW2/HW2.0/image.txt
