+++ MentoAD/HW2/HW2.1/HW2.1.txt
