--- MentoHomeWork/.idea/.idea.MentoHomeWork/.idea/projectSettingsUpdater.xml
@@ -1,6 +1,0 @@
-<?xml version="1.0" encoding="UTF-8"?>
-<project version="4">
-  <component name="RiderProjectSettingsUpdater">
-    <option name="vcsConfiguration" value="3" />
-  </component>
-</project>
